+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+## v0.15.0
+### feat
+- spell gem class and ethereal status, and description are available in battle screen now
+- add many spell gem UI screens
+  - refinery
+    - grind spell gem
+  - manage spell gem on your creature
+    - will speak class, whether it is the default, and speak when you change the creature page
+    - spell gem equip interface when managing spell gems
+  - enchanter
+    - craft
+    - enchant
+      - will speak number of properties and if you can't enchant anymore at the end of each enchant name on 1st UI change
+    - disenchant
+    - upgrade
+  - codex - spells
+
 ## v0.14.2
 ### feat
 - read Field Items UI creature selection